finished the dialog and send the image to the main window
This time, I finished the pop-up dialog which is abled to send the data
to the main window and the setted parameters in the dialog remain
unchanged compared with your last setting before the main window is
closed.
What's more, I use the signal slots to send the data to the main part
from the Qthread.