Feature test
The gray value read in the matlab is different from that in the openCV
in C++, the problem is still not found. Anyway, the priciple components
of the features in these two languages is similar, so this difference
between them is not important for the further analysis. Next work will
focus on the classifier part.